chore: rename incoming webhooks to signals
